fix: DeploymentService resourceConfig always empty — properly convert JsonNode to Map
Also fixes pre-existing compilation errors: BaseEntity missing @Id/@Setter,
DeploymentRepository/VersionRepository/AlertRuleRepository missing query methods,
ApprovalService ObjectMapper.builder() invalid API, InferenceService type mismatch.

+++ maidc-parent/maidc-model/src/main/java/com/maidc/model/service/InferenceService.java
@@ -58,7 +58,7 @@ public class InferenceService {
             return InferenceResultVO.builder()
                     .requestId(dto.getRequestId())
                     .results(dto.getInput()) // placeholder
-                    .latencyMs((int) latency)
+                    .latencyMs(latency)
                     .modelVersion("latest")
                     .build();
